Create new scSet object

Create new scSet object

createScSet(genome, protocol = NULL, tes = NULL, genes = NULL,
  reads = NULL, n_cores = NULL, seed = 19)

Arguments

genome
BSgenome object (required). Target genome assembly stored as BSgenome object.

tes
Path to repeatmasker output file or GRanges object with 'repname' metacolumn (required).

n_cores
Integer. Number of cores to use for downstream functions. If NULL (the default), detects the number of cores automatically. The doMC packge must be installed to register the cores.

seed
Integer. Seed for the random number generator. 19 by default.

alt_chromosomes
Logical. If FALSE (the default), restricts genome annotation to the main chromosome assembly.

cis
Path to bed file with cis regulatory feature coordinates or GRanges object (optional).

blacklist
Path to bed file with blacklisted regions or GRanges object (optional). Guides binding to blacklist regions are blacklisted.

whitelist
Path to bed file with whitelisted regions or GRanges object (optional). Guide off-target binding to whitelist regions are scored neutrally.

temp
Path to directory where temporary files are stored. Needs to be read- and writeable with sufficient storage space for large file sizes. If NULL (the default), is set to the return value of tempdir().

refdir
Path to search for bowtie index files. Will create new indeces in refdir if no corresponding files are found (i.e. do not match BSgenome prefix). If empty (the default), searches in the bowtie_indeces directory of the Repguide installation path.

Value

scSet object.

See also

Examples